Criado o relacionamento entre as tabelas e ajutado o editar.

+++ 4 - Desenvolvimento/database/migrations/2019_02_12_024740_create_distributors_table.php
@@ -16,9 +16,9 @@ class CreateDistributorsTable extends Migration
             $table->increments('id');
             $table->timestamps();
             $table->string('corporate_name')->nullable();
-            $table->integer('cnpj')->unsigned();
+            $table->double('cnpj')->unsigned();
             $table->text('address')->nullable();
-            $table->integer('phone')->unsigned();
+            $table->double('phone')->unsigned();
             $table->string('contact_person')->nullable();
             });
     }
